fix(ci): cross-platform + tool-availability guards for release multi-OS gate
- web: move TestServer_GracefulShutdownOnSIGTERM to !windows build-tag file (syscall.Kill undefined on windows — vet/compile failure)
- cli: skip TestAstgrepExitCode when sg (ast-grep) is not installed (CI runners lack sg; matches existing scanner_test.go skip pattern)

🗿 MoAI
